Move Default Inspector Header Config and some cleanup
